django-backend/requirement.txt
django
djangorestframework
PyJWT
bcrypt
mysqlclient
python-dotenv